+++ Practice/Assets/Script/IDamageable.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public interface IDamageable {
-    void TakeHit(float damage, Vector3 hitPoint, Vector3 hitDirection);
+    void TakeHit(int damage, Vector3 hitPoint, Vector3 hitDirection);
     void TakeHit(Color attackerColor, Vector3 hitPoint, Vector3 hitDirection, float knockbackForce);
-    void TakeDamage(float damage);
+    void TakeDamage(int damage);
 }
